update examples (#44)
* replace goerli w/ holesky in suave examples

* replace goerli w/ holesky in suave examples

* bun install

* chore: format

* update demos

* updating suave-std

* forge install: suave-std

* update suave-std

* update fn visibility in suave example to enable 'is Suapp'

* improve suave-web-demo readme

* fix comment

* fix suave example; make buildEthBlock work on all eth providers (requires suave-std upgrade)

* update suave-std (pin to commit da7ff46042139847cf32ba30afc1e24f4dc0bd86)

---------

+++ examples/suave/bids.ts
@@ -40,9 +40,9 @@ export class OFAOrder {
     return encodeAbiParameters([
       {
         components: [
-          { type: 'uint', name: 'blockNumber' },
-          { type: 'uint', name: 'minTimestamp' },
-          { type: 'uint', name: 'maxTimestamp' },
+          { type: 'uint64', name: 'blockNumber' },
+          { type: 'uint64', name: 'minTimestamp' },
+          { type: 'uint64', name: 'maxTimestamp' },
           { type: 'bytes[]', name: 'txns' },
         ],
         name: 'BundleObj',
